+++ crates/e2e/src/lib.rs
@@ -21,8 +21,9 @@
 
 mod backend;
 mod builders;
-mod client_utils;
+mod contract_build;
 mod contract_results;
+mod client_utils;
 mod drink_client;
 mod error;
 pub mod events;
@@ -30,6 +31,10 @@ mod node_proc;
 mod subxt_client;
 mod xts;
 
+pub use crate::contract_build::{
+    build_root_and_additional_contracts,
+    build_root_and_contract_dependencies,
+};
 pub use backend::{
     ChainBackend,
     ContractsBackend,